refactor: extract chunk logic for multiple queries
